password_manager_get_post/requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==43.0.1
Django==5.1.1
django-extensions==3.2.3
djangorestframework==3.15.2
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.2
pycparser==2.22
python-dotenv==1.0.1
sqlparse==0.5.1
